Avoid memory bank conflicts and use 32 bits offsets

+++ wavefront_structures.h
@@ -26,7 +26,7 @@
 
 #define SEQ_TYPE char
 
-typedef int16_t ewf_offset_t;
+typedef int32_t ewf_offset_t;
 typedef char edit_cigar_t;
 
 #define OFFSETS_TOTAL_ELEMENTS(max_d) (max_d * max_d + (max_d * 2 + 1))
